requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
aiohttp==3.11.2
azure-functions== 1.21.3
azure-functions-durable==1.2.10
azure-storage-blob==12.24.1
azure-identity==1.20.0
azure-keyvault-secrets==4.9.0
psycopg2-binary==2.9.10
backoff==2.2.1